refactor(ocr): split worker and word mapping helpers
Keep the OCR orchestration file below the project size guideline by moving reusable worker setup and OCR word coordinate mapping into domain-specific modules. This keeps future OCR fixes focused without changing runtime behavior.
