+++ src/core/ocr.ts
@@ -1,10 +1,18 @@
-import { readFile, writeFile } from 'node:fs/promises';
-import { createRequire } from 'node:module';
+import { readFile } from 'node:fs/promises';
 import { join } from 'node:path';
 import type { PDFDocumentProxy } from 'pdfjs-dist/legacy/build/pdf.mjs';
 import type { OcrWord, PageOcr } from '../types/index.js';
 import { ensurePrivateDir, getCacheRoot } from './cache.js';
-import { type RenderRegion, renderPageToBuffer, type ViewportCrop, viewportCropForRegion } from './renderer.js';
+import {
+  DEFAULT_OCR_RENDER_SCALE,
+  normaliseConfidence,
+  type OcrWordTransform,
+  transformOcrWords,
+} from './ocr/words.js';
+import { ensureQuietTesseractWorker } from './ocr/worker.js';
+import { type RenderRegion, renderPageToBuffer, viewportCropForRegion } from './renderer.js';
+
+export { buildQuietTesseractWorkerScript } from './ocr/worker.js';
 
 /**
  * One OCR worker, reusable across many pages. Created once per
@@ -20,162 +28,6 @@ export interface OcrSession {
 
 /** Lower bound of "this looks like a usable lang code" — letters only, 1+ chars. */
 const LANG_TOKEN = /^[A-Za-z_]+$/;
-const DEFAULT_RENDER_SCALE = 2;
-const QUIET_TESSERACT_WORKER_FILENAME = 'tesseract-quiet-worker.cjs';
-
-interface OcrWordTransform {
-  scale: number;
-  region?: RenderRegion;
-  crop?: ViewportCrop;
-  pageView?: readonly number[];
-  viewport?: PageViewportLike;
-}
-
-interface PageViewportLike {
-  convertToPdfPoint(x: number, y: number): number[];
-}
-
-interface PageBox {
-  x: number;
-  y: number;
-  width: number;
-  height: number;
-}
-
-interface RawOcrBbox {
-  x0: number;
-  y0: number;
-  x1: number;
-  y1: number;
-}
-
-interface RawOcrWord {
-  text?: unknown;
-  confidence?: unknown;
-  bbox?: RawOcrBbox;
-}
-
-function round2(n: number): number {
-  return Math.round(n * 100) / 100;
-}
-
-function round3(n: number): number {
-  return Math.round(n * 1000) / 1000;
-}
-
-function normaliseConfidence(value: unknown): number {
-  const raw = typeof value === 'number' ? value : 0;
-  return round3(Math.max(0, Math.min(1, raw / 100)));
-}
-
-function isUsableRawBbox(bbox: RawOcrBbox | undefined): bbox is RawOcrBbox {
-  return (
-    bbox !== undefined &&
-    Number.isFinite(bbox.x0) &&
-    Number.isFinite(bbox.y0) &&
-    Number.isFinite(bbox.x1) &&
-    Number.isFinite(bbox.y1) &&
-    bbox.x1 > bbox.x0 &&
-    bbox.y1 > bbox.y0
-  );
-}
-
-function isUsablePageView(value: readonly number[] | undefined): value is readonly [number, number, number, number] {
-  return Array.isArray(value) && value.length >= 4 && value.slice(0, 4).every((item) => Number.isFinite(item));
-}
-
-function isUsablePdfPoint(value: unknown): value is [number, number] {
-  return Array.isArray(value) && value.length >= 2 && Number.isFinite(value[0]) && Number.isFinite(value[1]);
-}
-
-function arrayProperty(value: unknown, key: string): unknown[] {
-  if (typeof value !== 'object' || value === null) return [];
-  const property = (value as Record<string, unknown>)[key];
-  return Array.isArray(property) ? property : [];
-}
-
-function collectRawWords(page: { blocks?: unknown }): RawOcrWord[] {
-  const out: RawOcrWord[] = [];
-  const blocks = arrayProperty(page, 'blocks');
-  for (const block of blocks) {
-    const paragraphs = arrayProperty(block, 'paragraphs');
-    for (const paragraph of paragraphs) {
-      const lines = arrayProperty(paragraph, 'lines');
-      for (const line of lines) {
-        const words = arrayProperty(line, 'words');
-        for (const word of words) out.push(word as RawOcrWord);
-      }
-    }
-  }
-  return out;
-}
-
-function ocrBboxToPageBox(bbox: RawOcrBbox, transform: OcrWordTransform): PageBox | undefined {
-  const scale = transform.scale > 0 ? transform.scale : DEFAULT_RENDER_SCALE;
-  const pageView = isUsablePageView(transform.pageView) ? transform.pageView : undefined;
-  const viewport = transform.viewport;
-  if (!viewport || !pageView) {
-    const offsetX = transform.region?.x ?? 0;
-    const offsetY = transform.region?.y ?? 0;
-    return {
-      x: round2(offsetX + bbox.x0 / scale),
-      y: round2(offsetY + bbox.y0 / scale),
-      width: round2((bbox.x1 - bbox.x0) / scale),
-      height: round2((bbox.y1 - bbox.y0) / scale),
-    };
-  }
-
-  const cropX = transform.crop?.x ?? 0;
-  const cropY = transform.crop?.y ?? 0;
-  const viewMinX = Math.min(pageView[0], pageView[2]);
-  const viewMaxY = Math.max(pageView[1], pageView[3]);
-  const corners = [
-    [bbox.x0, bbox.y0],
-    [bbox.x1, bbox.y0],
-    [bbox.x0, bbox.y1],
-    [bbox.x1, bbox.y1],
-  ];
-  const points = corners.map(([x, y]) => {
-    const pdfPoint = viewport.convertToPdfPoint(cropX + x, cropY + y);
-    if (!isUsablePdfPoint(pdfPoint)) return undefined;
-    const [pdfX, pdfY] = pdfPoint;
-    return {
-      x: pdfX - viewMinX,
-      y: viewMaxY - pdfY,
-    };
-  });
-  if (points.some((point) => point === undefined)) return undefined;
-  const usablePoints = points as { x: number; y: number }[];
-  const minX = Math.min(...usablePoints.map((point) => point.x));
-  const maxX = Math.max(...usablePoints.map((point) => point.x));
-  const minY = Math.min(...usablePoints.map((point) => point.y));
-  const maxY = Math.max(...usablePoints.map((point) => point.y));
-  return {
-    x: round2(minX),
-    y: round2(minY),
-    width: round2(maxX - minX),
-    height: round2(maxY - minY),
-  };
-}
-
-function transformOcrWords(page: { blocks?: unknown }, transform: OcrWordTransform): OcrWord[] {
-  const words: OcrWord[] = [];
-  for (const raw of collectRawWords(page)) {
-    const text = typeof raw.text === 'string' ? raw.text.trim() : '';
-    if (text.length === 0 || !isUsableRawBbox(raw.bbox)) continue;
-    const box = ocrBboxToPageBox(raw.bbox, transform);
-    if (!box) continue;
-    words.push({
-      text,
-      confidence: normaliseConfidence(raw.confidence),
-      x: box.x,
-      y: box.y,
-      width: box.width,
-      height: box.height,
-    });
-  }
-  return words;
-}
 
 /**
  * Parse a tesseract-style language string ("eng" / "eng+jpn" / "jpn+chi_sim")
@@ -196,27 +48,6 @@ export function parseOcrLang(lang: string): string[] {
     }
   }
   return tokens;
-}
-
-export function buildQuietTesseractWorkerScript(tesseractWorkerPath: string): string {
-  return `"use strict";
-const originalWrite = process.stderr.write.bind(process.stderr);
-const quiet = /^(?:Image too small to scale!! \\(\\d+x\\d+ vs min width of \\d+\\)|Line cannot be recognized!!)\\s*$/;
-process.stderr.write = (chunk, ...args) => {
-  const text = typeof chunk === "string" ? chunk : Buffer.isBuffer(chunk) ? chunk.toString("utf8") : String(chunk);
-  if (quiet.test(text)) return true;
-  return originalWrite(chunk, ...args);
-};
-require(${JSON.stringify(tesseractWorkerPath)});
-`;
-}
-
-async function ensureQuietTesseractWorker(cacheRoot: string): Promise<string> {
-  const requireFromHere = createRequire(import.meta.url);
-  const tesseractWorkerPath = requireFromHere.resolve('tesseract.js/src/worker-script/node/index.js');
-  const quietWorkerPath = join(cacheRoot, QUIET_TESSERACT_WORKER_FILENAME);
-  await writeFile(quietWorkerPath, buildQuietTesseractWorkerScript(tesseractWorkerPath), { mode: 0o600 });
-  return quietWorkerPath;
 }
 
 /**
@@ -330,7 +161,7 @@ export async function attachOcr(
       let png: Buffer;
       let contentRatio: number | undefined;
       const page = await doc.getPage(pageNumbers[i]);
-      const ocrScale = scale ?? DEFAULT_RENDER_SCALE;
+      const ocrScale = scale ?? DEFAULT_OCR_RENDER_SCALE;
       const viewport = page.getViewport({ scale: ocrScale });
       const transform: OcrWordTransform = {
         scale: ocrScale,
